tests/wpt/web-platform-tests/workers/constructors/SharedWorker/URLMismatchError.htm

<!DOCTYPE html>
<meta charset="utf-8">
<title>Web Workers: SharedWorker - same name, different URL</title>
<link rel="author" title="Intel" href="http://www.intel.com">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id="log"></div>
<script>
var counter = 0
function maybeDone() {
  if(counter) {
    done()
  }
  counter++
}

var worker = new SharedWorker('shared-worker.js', 'name');
worker.port.postMessage("trigger a response")
worker.port.onmessage = (e) => {
  assert_equals(e.data, "ping")
  maybeDone()
}

// This used to throw "URLMismatchError", but the standard changed
var worker2 = new SharedWorker('1', 'name');
worker2.port.onmessage = (e) => {
  assert_array_equals(e.data, ["1", "name"])
  maybeDone()
}
</script>
